Split tests in individual files
